Move package pkg/adjuster to internal/adjuster
