+++ internal/adjuster/bad_span_references.go
