+++ internal/adjuster/bad_span_references_test.go
