+++ internal/adjuster/doc.go
