+++ internal/storage/v1/memory/memory.go
@@ -14,9 +14,9 @@ import (
 	"github.com/gogo/protobuf/proto"
 
 	"github.com/jaegertracing/jaeger-idl/model/v1"
+	"github.com/jaegertracing/jaeger/internal/adjuster"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/spanstore"
 	"github.com/jaegertracing/jaeger/internal/tenancy"
-	"github.com/jaegertracing/jaeger/model/adjuster"
 )
 
 // Store is an in-memory store of traces
